更新了测试代码

+++ YohunlCategory/NYXImagesKit-master/Categories/UIImage+Filtering.m
@@ -693,7 +693,7 @@ static float __f_unsharpen_kernel_3x3[9] = {
 		void* outt = malloc(n);
 		vImage_Buffer src = {data, height, width, bytesPerRow};
 		vImage_Buffer dest = {outt, height, width, bytesPerRow};
-		vImageConvolveWithBias_ARGB8888(&src, &dest, NULL, 0, 0, __s_sharpen_kernel_3x3, 3, 3, 1/*divisor*/, bias, NULL, kvImageCopyInPlace);
+		vImageConvolveWithBias_ARGB8888(&src, &dest, NULL, 0, 0, __s_sharpen_kernel_3x3, 3, 3, 1/*divisor*/, (int32_t)bias, NULL, kvImageCopyInPlace);
 
 		memcpy(data, outt, n);
